BRUUUUUUUUUUUUUUUHHHH
Prince Vasily did not think over his plans, he even less thought to do evil to people in order to gain profit. He was only a secular person who succeeded in the world and made a habit out of this success. He constantly, depending on the circumstances, according to the rapprochement with people, various plans and considerations were drawn up, in which he himself did not give himself a good account, but which constituted the whole interest of his life. Not one or two such plans and considerations were in use for him, but dozens of which some were just beginning to appear to him, others were achieved, and still others were destroyed. The Countess was so tired of the visits that she did not order anyone else to be received, and the doorman was only ordered to invite everyone who would still come with congratulations to eat. The Countess wanted to talk privately with a friend about her childhood, Princess Anna Mikhailovna, whom she had not seen well since her arrival from St. Petersburg. Anna Mikhailovna, with her tear-stained and pleasant face, moved closer to the Countess's chair. “I will be completely frank with you,” said Anna Mikhailovna. - There are too few of us, old friends! That is why I value your friendship so much.
Anna Mikhailovna looked at Vera and stopped. The Countess shook hands with her friend.
“Vera,” said the Countess, addressing her eldest daughter, who was obviously unloved. - How do you have no idea about anything? Don't you feel like you're superfluous here? Go to your sisters, or ...
The beautiful Vera smiled contemptuously, apparently not feeling the slightest insult.